+++ modules/core/dependency/python-ihm/ihm/reader.py
@@ -1495,6 +1495,8 @@ class _DatasetListHandler(Handler):
             (x[1].data_type.lower(), x[1])
             for x in inspect.getmembers(ihm.dataset, inspect.isclass)
             if issubclass(x[1], ihm.dataset.Dataset))
+        # Map old 'CX-MS' data to new class
+        self.type_map['cx-ms data'] = ihm.dataset.CXMSDataset
 
     def __call__(self, data_type, id, details):
         typ = None if data_type is None else data_type.lower()
@@ -1721,8 +1723,8 @@ class _StartingComparativeModelsHandler(Handler):
         dataset = self.sysr.datasets.get_by_id(template_dataset_list_id)
         aln = self.sysr.external_files.get_by_id_or_none(alignment_file_id)
         asym_id = template_auth_asym_id
-        seq_id_range = (int(starting_model_seq_id_begin),
-                        int(starting_model_seq_id_end))
+        seq_id_range = (self.get_int(starting_model_seq_id_begin),
+                        self.get_int(starting_model_seq_id_end))
         template_seq_id_range = (self.get_int(template_seq_id_begin),
                                  self.get_int(template_seq_id_end))
         identity = ihm.startmodel.SequenceIdentity(
@@ -2947,8 +2949,8 @@ class _CrossLinkResultHandler(Handler):
             sigma2=self.get_float(sigma_2))
 
 
-class _OrderedEnsembleHandler(Handler):
-    category = '_ihm_ordered_ensemble'
+class _OrderedModelHandler(Handler):
+    category = '_ihm_ordered_model'
 
     def __call__(self, process_id, step_id, model_group_id_begin,
                  model_group_id_end, edge_description, ordered_by,
@@ -2971,6 +2973,12 @@ class _OrderedEnsembleHandler(Handler):
             mapkeys={'process_description': 'description'})
         self.copy_if_present(
             step, locals(), mapkeys={'step_description': 'description'})
+
+
+# Handle the old name for the ihm_ordered_model category. This is a separate
+# object so relies on _OrderedModelHandler not storing any state.
+class _OrderedEnsembleHandler(_OrderedModelHandler):
+    category = '_ihm_ordered_ensemble'
 
 
 class UnknownCategoryWarning(Warning):
@@ -3801,7 +3809,7 @@ class IHMVariant(Variant):
         _BranchDescriptorHandler, _BranchLinkHandler, _CrossLinkListHandler,
         _CrossLinkRestraintHandler, _CrossLinkPseudoSiteHandler,
         _CrossLinkResultHandler, _StartingModelSeqDifHandler,
-        _OrderedEnsembleHandler,
+        _OrderedModelHandler, _OrderedEnsembleHandler,
         _MultiStateSchemeHandler, _MultiStateSchemeConnectivityHandler,
         _KineticRateHandler,
         _RelaxationTimeHandler, _RelaxationTimeMultiStateSchemeHandler