+++ modules/core/dependency/python-ihm/test/test_location.py
@@ -114,6 +114,15 @@ class Tests(unittest.TestCase):
         self.assertEqual(d.version, 1)
         self.assertEqual(d.details, 'foo')
 
+    def test_proteome_xchange_location(self):
+        """Test ProteomeXchangeLocation class"""
+        d = ihm.location.ProteomeXchangeLocation(
+            'abc', version=1, details='foo')
+        self.assertEqual(d.db_name, 'ProteomeXchange')
+        self.assertEqual(d.access_code, 'abc')
+        self.assertEqual(d.version, 1)
+        self.assertEqual(d.details, 'foo')
+
     def test_empiar_location(self):
         """Test EMPIARLocation class"""
         d = ihm.location.EMPIARLocation('abc', version=1, details='foo')